feat(.pre-commit-config.yaml): add pre-commit hooks configuration
This commit adds the configuration for pre-commit hooks in the new file `.pre-commit-config.yaml`. The configuration includes several hooks from the `pre-commit/pre-commit-hooks` repository, such as `check-yaml`, `end-of-file-fixer`, `trailing-whitespace`, `check-symlinks`, `destroyed-symlinks`, `detect-aws-credentials`, `check-shebang-scripts-are-executable`, `check-merge-conflict`, and `detect-private-key`. Additionally, a hook from the `IamTheFij/docker-pre-commit` repository, `docker-compose-check`, is also included in the configuration.

Signed-off-by: Rodney Osodo <28790446+rodneyosodo@users.noreply.github.com>