Refactor compose.sh script and remove unnecessary code
The changes in this commit involve refactoring the compose.sh script. Specifically, the changes remove unnecessary code related to the docker-compose command. The previous command "docker-compose" has been replaced with "docker compose" for both the "start" and "stop" options. Additionally, the generation of the docker-compose file for the "code-server" service has been removed from the yaml file.

Signed-off-by: Rodney Osodo <28790446+rodneyosodo@users.noreply.github.com>