feat(docker-compose): Add swagger-editor service
This commit adds the swagger-editor service to the docker-compose.yaml file. The service is responsible for running the swaggerapi/swagger-editor:v4.12.1 image with the specified configurations. The service is named swagger-editor and is set to restart unless stopped. It is connected to the homelab-network and exposes port 2080, which maps to port 8080 inside the container. Additionally, the security_opt option is set to no-new-privileges:true.

Signed-off-by: Rodney Osodo <28790446+rodneyosodo@users.noreply.github.com>