feat(pi-hole): update IP addresses and dependencies
Updated the IP addresses for the Pi-hole and Unbound services in the Docker Compose file. The Pi-hole IP address was changed from 172.30.0.10 to 172.30.0.100, and the Unbound IP address was changed from 172.30.0.11 to 172.30.0.101. Additionally, added a dependency for the Pi-hole service to depend on the Unbound service. These changes ensure proper network configuration and functionality between the services.

Signed-off-by: Rodney Osodo <28790446+rodneyosodo@users.noreply.github.com>