feat(opengist): Add opengist service to docker-compose
Add the opengist service to the docker-compose.yaml file in the opengist directory. This service allows for hosting and managing gists using the opengist image. The service is configured to run on port 3020 and has a volume mounted for data persistence. Additionally, the service is assigned the UID and GID values of 1000 for proper permissions.

Signed-off-by: Rodney Osodo <28790446+rodneyosodo@users.noreply.github.com>