feat(docker-compose): Add dozzle service to docker-compose.yaml
This commit adds the dozzle service to the docker-compose.yaml file. The dozzle service uses the amir20/dozzle:v6.1.1 image and is configured to restart unless stopped. It is connected to the homelab-network and has the security option no-new-privileges set to true. Additionally, it mounts the /var/run/docker.sock as a read-only volume and exposes port 3070 to port 8080.

Signed-off-by: Rodney Osodo <28790446+rodneyosodo@users.noreply.github.com>