feat: Add "mealie" service and modify "nextcloud"
The changes made to the docker-compose.yaml file include the addition of a new service called "mealie" and modifications to the "nextcloud" service. The "mealie" service has been added to enhance the functionality of the application. The "nextcloud" service has been modified to update the SMTP settings and trusted domains, improving the email functionality and security of the application.

Signed-off-by: Rodney Osodo <28790446+rodneyosodo@users.noreply.github.com>