feat: Add Hedgedoc service to docker-compose
Added a new service for Hedgedoc in the docker-compose file. The service includes the necessary configurations and dependencies for running Hedgedoc. The service uses the latest version of the Hedgedoc image (1.9.9) and is set to restart unless stopped. The service is connected to the homelab-network and has the required environment variables for connecting to the Hedgedoc database. The service also includes volume mappings for storing uploads and uses port 3050 for communication. The Hedgedoc service depends on the hedgedoc-database service, which runs a PostgreSQL database. The database is configured with the necessary environment variables and volume mappings. Both services have the necessary security options enabled.

Signed-off-by: Rodney Osodo <28790446+rodneyosodo@users.noreply.github.com>