feat(docker-compose): Add Rallly service to docker-compose
Added a new service called Rallly to the docker-compose.yaml file. This service runs the Rallly application using the lukevella/rallly:3.5.1 image. It is configured to restart unless stopped and is connected to the homelab-network. The Rallly service depends on the rallly-db service and exposes port 3060. Various environment variables are set for the Rallly service, including the database URL, secret password, base URL, allowed emails, and SMTP configuration. The rallly-db service is also added, which uses the postgres:16.1 image and is connected to the homelab-network. The service mounts a volume for persistent data storage and sets environment variables for the Postgres user, password, and database. A healthcheck is configured for the rallly-db service to ensure its availability.

Signed-off-by: Rodney Osodo <28790446+rodneyosodo@users.noreply.github.com>