feat: Add hemmelig service to docker-compose.yaml
This commit adds a new service called "hemmelig" to the docker-compose.yaml file. The "hemmelig" service uses the image "hemmeligapp/hemmelig:v5.19.4" and is set to restart unless stopped. It is connected to the "homelab-network" and has the security option "no-new-privileges:true". The service also has environment variables and volumes configured for data storage. Additionally, it is exposed on port 3080 and has a healthcheck configured.

Signed-off-by: Rodney Osodo <28790446+rodneyosodo@users.noreply.github.com>